+++ src/universal/index.mjs
@@ -1,0 +1,1 @@
+export { withGraphQLApp } from './withGraphQLApp'