+++ src/universal/ssrAlias.mjs
@@ -1,0 +1,6 @@
+/**
+ * An empty named export used by [`withGraphQLConfig`]{@link withGraphQLConfig}
+ * to alias the server only module `graphql-react/lib/ssr` for a client bundle.
+ * @ignore
+ */
+export let ssr